fastapi-backend/test_enhanced_image_service.py
"""
Test script for the enhanced image generation service
"""
import asyncio
import os
from dotenv import load_dotenv
from services.enhanced_image_service import enhanced_image_service
from services.presentation_db_service import presentation_db_service

# Load environment variables
load_dotenv()

async def test_image_generation():
    """Test the DALL-E image generation and GCS upload"""
    print("🚀 Testing DALL-E image generation with GCS upload...")
    
    try:
        # Test image generation
        prompt = "modern office workspace with computers and plants, professional lighting"
        print(f"📝 Generating image for prompt: '{prompt}'")
        
        image_url = await enhanced_image_service.generate_presentation_image(prompt)
        print(f"✅ Image generated successfully!")
        print(f"🔗 Image URL: {image_url}")
        
        # Test database save (create a test user first)
        user_email = "test@example.com"
        print(f"💾 Saving image metadata to database for user: {user_email}")
        
        # Note: We'll skip the database save for now due to user relationship
        # saved_image = await presentation_db_service.save_generated_image(
        #     url=image_url,
        #     prompt=prompt,
        #     user_email=user_email,
        #     model="dall-e-3"
        # )
        # print(f"✅ Image metadata saved to database: {saved_image}")
        
        print("🎉 All tests passed!")
        return True
        
    except Exception as e:
        print(f"❌ Error during testing: {e}")
        import traceback
        traceback.print_exc()
        return False

if __name__ == "__main__":
    success = asyncio.run(test_image_generation())
    if success:
        print("\n✅ Enhanced image service is working correctly!")
        print("🔄 Ready to integrate with FastAPI presentation router")
    else:
        print("\n❌ Tests failed. Please check the error messages above.")
