fastapi-backend/requirements.txt
fastapi
fastapi[standard]
uvicorn[standard]
langchain
langchain-openai
pydantic
python-dotenv
google-cloud-storage
python-multipart
openai
prisma
requests
httpx
python-docx
reportlab
Pillow
asyncio-throttle
